Accept fractional zone minutes (Apple Health uses decimals)
ZoneTarget.minutes was declared u32, but Apple Health samples HR
continuously and reports fractional zone time — e.g. an Elliptical
session as [{Z1, 2.45}, {Z2, 17.85}, {Z3, 2.65}, {Z4, 15.6}, {Z5, 1.07}].
parse_cardio_actuals couldn't parse 2.45 as u32, fell back to the bare-
object parse path (which has no top-level "zones" because the input is
wrapped), and surfaced the misleading "JSON parse: missing field `zones`"
error in the UI.

Two fixes:

1. ZoneTarget.minutes: u32 → f32. Apple Health zone times are now
   represented honestly. Legacy state.json files with integer literals
   like `"minutes": 30` still deserialize cleanly — serde reads integers
   into floats without complaint. Consumers updated:
   - coach.rs cardio_minutes_in_window sums per-set zones as f32 and
     rounds once before adding to the u32 weekly total (avoids
     accumulated truncation bias)
   - components/exercises.rs apply_cardio_actuals_to_freeform sums
     zones as f32, rounds to u32 to seed set.reps
   - components/session.rs zone editor parses input as f32

2. parse_cardio_actuals error reporting: when both wrapped and bare
   parses fail, prefer the wrapped error if the body clearly contains
   "cardio_actuals" at the top level. The bare-form fallback's "missing
   field" error was hiding the actual underlying problem.

Tests (4 new, all green):
- cardio_actuals_accepts_fractional_minutes — the user's exact bug
  payload, asserting sum = 39.62
- cardio_actuals_legacy_integer_minutes_still_parse — back-compat for
  pre-existing state.json files
- cardio_actuals_error_for_wrapped_form_reports_underlying_problem —
  bad nested type should surface "invalid type", not "missing field zones"
- cardio_minutes_in_window_sums_fractional_zones_and_rounds — 39.62
  rounds to 40 for the weekly u32 total

+++ src/models.rs
@@ -184,10 +184,15 @@ pub enum WorkoutSource {
 /// Apple-Health-style HR zone budget. Used both as a coach prescription
 /// (`target_zones`) and as a per-set actual (`zone_minutes`). Zone 1-5 maps
 /// to the standard Apple Watch HR zones; we don't try to redefine them.
+///
+/// `minutes` is `f32`, not `u32`, because Apple Health samples HR continuously
+/// and reports fractional zone time (e.g. `17.85` min in Z2). Legacy JSON
+/// with integer values like `"minutes": 30` still deserializes — serde reads
+/// integers into floats without complaint.
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct ZoneTarget {
     pub zone: u8,
-    pub minutes: u32,
+    pub minutes: f32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
